body calculator

+++ app/src/main/java/com/progressor/progressor/model/dataobjects/account/Body.kt
@@ -3,10 +3,10 @@ package com.progressor.progressor.model.dataobjects.account
 import java.time.LocalDateTime
 
 data class Body(
-        val weight: Int? = null,
-        val fatPercentage: Int? = null,
-        val musclePercentage: Int? = null,
-        val waterPerfentage: Int? = null,
+        val weight: Double? = null,
+        val fatPercentage: Double? = null,
+        val musclePercentage: Double? = null,
+        val waterPerfentage: Double? = null,
         val mood: String? = null,
         val createDate: LocalDateTime? = null
 )
